Move `Analyst` to file for its generating endpoint
