+++ lib/reusable/lovat_api/get_analysts.dart
@@ -1,6 +1,7 @@
 import 'dart:convert';
 
 import 'package:scouting_dashboard_app/reusable/lovat_api/lovat_api.dart';
+import 'package:scouting_dashboard_app/reusable/lovat_api/promote_analyst.dart';
 
 extension GetAnalysts on LovatAPI {
   Future<List<Analyst>?> getAnalysts() async {
@@ -17,5 +18,33 @@ extension GetAnalysts on LovatAPI {
     final json = jsonDecode(response!.body) as List<dynamic>;
 
     return json.map((e) => Analyst.fromJson(e)).toList();
+  }
+}
+
+class Analyst {
+  const Analyst({
+    required this.id,
+    required this.name,
+    required this.email,
+  });
+
+  final String id;
+  final String name;
+  final String email;
+
+  factory Analyst.fromJson(Map<String, dynamic> json) => Analyst(
+        id: json['id'] as String,
+        name: json['username'] as String,
+        email: json['email'] as String,
+      );
+
+  Map<String, dynamic> toJson() => {
+        'id': id,
+        'username': name,
+        'email': email,
+      };
+
+  Future<void> promote() async {
+    await lovatAPI.promoteAnalyst(id);
   }
 }
